chore: update api-proxy project structure
