+++ api-proxy/src/handlers/mod.rs
@@ -1,0 +1,3 @@
+pub mod health_handlers;
+pub mod block_handlers;
+pub mod custom_handlers;
