+++ api-proxy/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod block_service;
+
